Initial replacement of VertexArrays with GL3 RenderBuffers (easily GL4-able later). (#141)
* Removed VertexArray.inl
* Semantics changes in StreamBuffer. Added manual command to SwapBuffer(). Added support for partial mapping
* RenderBuffers, a (more) modern GL counterpart of VertexArray
* Added RenderBuffers QuadConverters (WIP)
* Added semi-templated RenderBuffer shaders
* Remove old VertexArrayRange
* Add ability to reload text only (no files) shaders
* Moved VA_TYPE_X to rts/Rendering/GL/VertexArrayTypes.h
* Converted CSelectedUnitsHandler::Draw() to RenderBuffers (WIP)
* Ported develop's Rendering/Fonts/*
* Fix StreamBuffer and RenderBuffer issues
* Fix FixedPipelineState. Still pretty bad design and crashes on MinGW. Remove?
* Ported develop ProfileDrawer.h/ProfileDrawer.cpp
* Changed TimeProfiler mutex type.

+++ rts/Sim/Projectiles/ExplosionGenerator.cpp
@@ -732,7 +732,7 @@ void CCustomExplosionGenerator::ParseExplosionCode(
 		if (!isFloat || memberInfo.length < 3)
 			throw content_error("[CCEG::ParseExplosionCode] incorrect use of \"dir\" (" + script + ")");
 
-		const std::uint16_t ofs = memberInfo.offset;
+		const std::uint16_t ofs = static_cast<uint16_t>(memberInfo.offset);
 
 		code.append(1, OP_DIR);
 		code.append((char*) &ofs, (char*) &ofs + sizeof(ofs));
@@ -765,7 +765,7 @@ void CCustomExplosionGenerator::ParseExplosionCode(
 		code.append(1, OP_LOADP);
 		code.append((char*)(&ptr), ((char*)(&ptr)) + sizeof(void*));
 
-		const std::uint16_t ofs = memberInfo.offset;
+		const std::uint16_t ofs = static_cast<uint16_t>(memberInfo.offset);
 
 		code.append(1, OP_STOREP);
 		code.append((char*)&ofs, (char*)&ofs + sizeof(ofs));
@@ -840,7 +840,7 @@ void CCustomExplosionGenerator::ParseExplosionCode(
 	}
 
 	// store the final value
-	const std::uint16_t ofs = memberInfo.offset;
+	const std::uint16_t ofs = static_cast<uint16_t>(memberInfo.offset);
 
 	code.push_back(isFloat ? OP_STOREF : OP_STOREI);
 	code.push_back(memberInfo.size);
